+++ packages/polyglot/src/Embeddings/Config/EmbeddingsRetryPolicy.php
@@ -50,7 +50,7 @@ final readonly class EmbeddingsRetryPolicy
     public function delayMsForAttempt(int $attemptNumber): int {
         $attempt = max(1, $attemptNumber);
         $base = $this->baseDelayMs * (2 ** ($attempt - 1));
-        $capped = min($base, $this->maxDelayMs);
+        $capped = (int) min($base, $this->maxDelayMs);
 
         return match ($this->jitter) {
             'none' => $capped,
